Fix GitCommit.ts type (#16)
* bump api to 0.1.4

* use int64 for GitCommit.ts field

* fix ts type

* generate rust client

+++ clients/rust/src/git_commit.rs
@@ -8,11 +8,11 @@ pub struct GitCommit {
     #[serde(rename="message")]
     pub message: String,
     #[serde(rename="ts")]
-    pub ts: i32,
+    pub ts: i64,
 }
 
 impl GitCommit {
-    pub fn new(oid: String, header: String, message: String, ts: i32) -> GitCommit {
+    pub fn new(oid: String, header: String, message: String, ts: i64) -> GitCommit {
         GitCommit {
             oid,
             header,
